requirements.txt
accelerate==1.12.0
aenum==3.1.16
annotated-types==0.7.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
botocore==1.42.30
Cerberus==1.3.8
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
invoke==2.2.1
islpy==2025.2.5
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnx==1.17.0
onnx-graphsurgeon==0.3.27
onnx-ir==0.1.14
onnxruntime==1.23.2
onnxscript==0.5.7
onnxsim==0.4.36
packaging==25.0
pandas==2.3.3
paramiko==4.0.0
pillow==12.1.0
prettytable==3.17.0
protobuf==3.20.3
psutil==7.2.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
qai-hub==0.42.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
s3transfer==0.13.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.1
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
timm==1.0.24
tokenizers==0.22.2
torch==2.9.1+cpu
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
